refactor(captage): extract DualRangeSlider into its own file
- Move DualRangeSlider to dual-range-slider.tsx (190 lines)
- captage-analyses-header.tsx reduced from 617 to 428 lines
- Remove orphan section comment
